UT_GenericVector<> now use UT_sint32 consistently to eliminate a lot of warnings
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@25101 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ plugins/collab/backends/service/xp/RealmProtocol.cpp
@@ -99,7 +99,7 @@ int RoutingPacket::parse(const char* buf, size_t size) {
 		return -1;
 	// get the recipients
 	m_address_count = buf[parsed];
-	if (m_address_count + 1 > getPayloadSize())
+	if ((uint32_t)m_address_count + 1 > getPayloadSize())
 		return -1;
 	m_connection_ids.resize(m_address_count);
 	std::copy(buf+parsed+1, buf+parsed+1+m_address_count, m_connection_ids.begin());
